boxoffice/requirements.txt
flask
pymysql
python-dotenv
cryptography
